% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootViz.R
\name{bootViz}
\alias{bootViz}
\title{Produce visualizations for boostrap results}
\usage{
bootViz(.result = NULL, type = c("histogram", "density"), bins = 30)
}
\arguments{
\item{.result}{A matrix returned from CalcGTheoryCI.}

\item{type}{Visualization produced: histogram or density plot. Defaults to histogram.}

\item{bins}{The number of bins for histograms}
}
\value{
A ggplot object containing small multiples.
}
\description{
This function creates histograms or density plots for results from CalcGTheoryCI
}
\examples{
\dontrun{
data("Brennan.3.1")
set.seed(123)
test_boot <- CalcGTheoryCI(Data = Brennan.3.1, B = 1000, type = "pi")
bootViz(test_boot, type = "density")
}
}
