% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Brennan.3.1}
\alias{Brennan.3.1}
\title{Data from Brennan}
\format{
A data frame with 80 rows and 4 variables:
\describe{
\item{Person}{Person ID}
\item{Item}{Item ID}
\item{Occasion}{Occasion ID}
\item{Score}{Given score}
}
}
\usage{
Brennan.3.1
}
\description{
A dataset containing 4 variables: Person, Item, Occasion, & Score
}
\keyword{datasets}
