% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdjustedVar.R
\name{calcAdjustedVar}
\alias{calcAdjustedVar}
\title{Calculate adjusted variance components.}
\usage{
calcAdjustedVar(
  .data,
  .var_p,
  .var_i,
  .var_o,
  .var_pi,
  .var_po,
  .var_io,
  .var_pio,
  type
)
}
\arguments{
\item{.data}{Internal input}

\item{.var_p}{Internal input}

\item{.var_i}{Internal input}

\item{.var_o}{Internal input}

\item{.var_pi}{Internal input}

\item{.var_po}{Internal input}

\item{.var_io}{Internal input}

\item{.var_pio}{Internal input}

\item{type}{Internal input}
}
\description{
This is an internal function for CalcGTheoryCI
}
