% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdjustedVarNested.R
\name{calcAdjustedVarNested}
\alias{calcAdjustedVarNested}
\title{Calculate adjusted variance components for the nested model.}
\usage{
calcAdjustedVarNested(.data, .var_p, .var_o, .var_po, .var_io, .var_pio, type)
}
\arguments{
\item{.data}{Internal input}

\item{.var_p}{Internal input}

\item{.var_o}{Internal input}

\item{.var_po}{Internal input}

\item{.var_io}{Internal input}

\item{.var_pio}{Internal input}

\item{type}{Internal input}
}
\description{
This is an internal function for CalcGTheoryCINested
}
