% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGTheoryCINested.R
\name{CalcGTheoryCINested}
\alias{CalcGTheoryCINested}
\title{Compute confidence intervals for G-theory quantities for the nested model}
\usage{
CalcGTheoryCINested(
  Data = NULL,
  B = 1000,
  type = c("p", "o", "po", "io", "pio")
)
}
\arguments{
\item{Data}{Data in long format in the following order: person, facet1, facet2 (facet1 is nested within facet2), score}

\item{B}{Number of bootstrap iterations}

\item{type}{Type of bootstrap sampling scheme as described in Tong & Brennan (2004)}
}
\value{
A matrix of bootstrap values which is then given to the summaryCI() function
}
\description{
This function computes and stores the raw bootstrap samples for G-theory variance components and coefficients.
}
\examples{
\dontrun{
data("Brennan.3.1")
CalcGTheoryCINested(Data = Brennan.3.1, B = 1000, type = "po")
}
}
