% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVarComp.R
\name{calcVarComp}
\alias{calcVarComp}
\title{Calculate variance components.}
\usage{
calcVarComp(Data = NULL)
}
\arguments{
\item{Data}{Data in long format in the following order: person, facet1, facet2, score}
}
\value{
A list of variance components
}
\description{
This function computes 2-facet crossed variance components using ANOVA.
}
\examples{
\dontrun{
data("Brennan.3.1")
calcVarComp(Data = Brennan.3.1)
}
}
