% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVarCompNested.R
\name{calcVarCompNested}
\alias{calcVarCompNested}
\title{Calculate nested variance components.}
\usage{
calcVarCompNested(Data = NULL)
}
\arguments{
\item{Data}{Data in long format in the following order: person, facet1, facet2 (facet1 is nested within facet2), score}
}
\value{
A list of variance components
}
\description{
This function computes 2-facet nested variance components using ANOVA.
}
\examples{
\dontrun{
data("Brennan.3.1")
calcVarCompNested(Data = Brennan.3.1)
}
}
