% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_DesMat_2crossed.R
\name{construct_DesMat_2crossed}
\alias{construct_DesMat_2crossed}
\title{Create linear random effects model design matrix for simulated data}
\usage{
construct_DesMat_2crossed(np, ni, nr)
}
\arguments{
\item{np}{Sample size for persons}

\item{ni}{Sampe size for facet 1}

\item{nr}{Sample size for facet 2}
}
\value{
A design matrix which is then given to the sim_study_2crossed() function
}
\description{
This function produces a linear random effects model design matrix for simulating data from a two-facet crossed design.
}
