% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinySimulation.R
\name{runShinySimulation}
\alias{runShinySimulation}
\title{Reactive Gboot Simulations with Shiny}
\usage{
runShinySimulation()
}
\value{
Runs a shiny app in your local session
}
\description{
This function will let you tweak simulated data parameters and view results
}
\examples{
\dontrun{
runShinySimulation()
}
}
