% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_2crossed.R
\name{sim_dat_2crossed}
\alias{sim_dat_2crossed}
\title{Generates simulated data set for two-facet crossed design.}
\usage{
sim_dat_2crossed(
  np,
  ni,
  nr,
  sig_p,
  sig_i,
  sig_r,
  sig_pi,
  sig_pr,
  sig_ir,
  sig_pir
)
}
\arguments{
\item{np}{Sample size for persons for simulated data.}

\item{ni}{Sample size for Facet 1 for simulated data.}

\item{nr}{Sample size for Facet 2 for simulated data.}

\item{sig_p}{Generating (i.e. population) value for person standard deviation.}

\item{sig_i}{Generating (i.e. population) value for Facet 1 deviation.}

\item{sig_r}{Generating (i.e. population) value for Facet 2 deviation.}

\item{sig_pi}{Generating (i.e. population) value for person X Facet 1 interaction standard deviation.}

\item{sig_pr}{Generating (i.e. population) value for person X Facet 2 interaction standard deviation.}

\item{sig_ir}{Generating (i.e. population) value for Facet 1 X Facet 2 standard deviation.}

\item{sig_pir}{Generating (i.e. population) value for residual standard deviation.}
}
\value{
A data frame in long format with columns Person, Facet 1, Facet 2, and simulated Scores
}
\description{
This function generates simulated data for the two-facet crossed design.
}
\examples{
\dontrun{
#Generate a single data set with sample sizes and parameters as in Table 3 of Tong and Brennan (2004)
sim_dat_2crossed(np = 100, ni = 20, nr = 2,
                 sig_p = 4, sig_i =  2, sig_r = 1,
                 sig_pi = 8, sig_pr = sqrt(2), sig_ir = sqrt(3), sig_pir = 12)

}
}
\references{
Tong, Y., & Brennan, R. L. (2004). Bootstrap techniques for estimating variability in generalizability theory. (Report No. 15).  Center for Advanced Studies in Measurement and Assessment, University of Iowa.
}
