% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCI.R
\name{summaryCI}
\alias{summaryCI}
\title{Produces summary confidence intervals from CalcGTheoryCI function.}
\usage{
summaryCI(.result = NULL, ConfLevel = 0.8, rounding = 4)
}
\arguments{
\item{.result}{A matrix returned from CalcGTheoryCI.}

\item{ConfLevel}{The confidence level. Should be between 0 and 1, Defaults to .8.}

\item{rounding}{The number of rounding digits. Defaults to 4}
}
\value{
A list of matrices.
}
\description{
Produces summary confidence intervals from CalcGTheoryCI function.
}
\examples{
\dontrun{
data("Brennan.3.1")
set.seed(123)
test_boot <- CalcGTheoryCI(Data = Brennan.3.1, B = 1000, type = "pi")
sumTest <- summaryCI(test_boot, 0.80, 4)
}
}
