% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryCINested.R
\name{summaryCINested}
\alias{summaryCINested}
\title{Produces summary confidence intervals from CalcGTheoryCINested function.}
\usage{
summaryCINested(.result = NULL, ConfLevel = 0.8, rounding = 4)
}
\arguments{
\item{.result}{A matrix returned from CalcGTheoryCINested.}

\item{ConfLevel}{The confidence level. Should be between 0 and 1, Defaults to .8.}

\item{rounding}{The number of rounding digits. Defaults to 4}
}
\value{
A list of matrices.
}
\description{
Produces summary confidence intervals from CalcGTheoryCINested function.
}
\examples{
\dontrun{
data("Brennan.3.1")
set.seed(123)
test_boot <- CalcGTheoryCINested(Data = Brennan.3.1, B = 1000, type = "po")
sumTest <- summaryCINested(test_boot, 0.80, 4)
}
}
