% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGTheoryCI.R
\name{CalcGTheoryCI}
\alias{CalcGTheoryCI}
\title{Compute confidence intervals for G-theory quantities}
\usage{
CalcGTheoryCI(Data, B = 1000, type)
}
\arguments{
\item{Data}{Data in long format in the following order: person, facet1, facet2, score}

\item{B}{Number of bootstrap iterations}

\item{type}{Type of bootstrap sampling scheme as desribed in Tong & Brennan (2004)}
}
\value{
A matrix of bootstrap values which is then given to the summaryCI() function
}
\description{
This function computes and stores the raw bootstrap samples for G-theory variance components and coefficients.
}
